Move actioner_models to configs.actioner
