+++ hasher-matcher-actioner/hmalib/lambdas/actions/action_performer.py
@@ -7,9 +7,7 @@ from functools import lru_cache
 from mypy_boto3_dynamodb import DynamoDBServiceResource
 
 from hmalib.common.message_models import ActionMessage
-from hmalib.common.actioner_models import (
-    ActionPerformer,
-)
+from hmalib.common.configs.actioner import ActionPerformer
 from hmalib.common.content_models import ActionEvent
 from hmalib.common.logging import get_logger
 from hmalib.common import config
